Add ShowDetails and SeasonDetails view.
