+++ ios/ios/UI/SeasonDetails/SeasonDetailsView.swift
@@ -9,54 +9,69 @@
 import SwiftUI
 import SwiftUIComponents
 import TvManiac
-import TvManiacUI
 import TvManiacKit
+import TvManiacUI
 
 struct SeasonDetailsView: View {
   private let presenter: SeasonDetailsPresenter
 
   @Environment(\.presentationMode) var presentationMode
-    
+
   @StateFlow private var uiState: SeasonDetailState
   @State private var isTruncated = false
   @State private var showFullText = false
   @State private var showModal = false
-  @State private var scrollOffset: CGFloat = 0
-    
+  @State private var showGlass: Double = 0
+  @State private var progressViewOffset: CGFloat = 0
+
   init(presenter: SeasonDetailsPresenter) {
     self.presenter = presenter
     _uiState = StateFlow(presenter.state)
   }
-    
+
   var body: some View {
     ZStack {
       Color.background.edgesIgnoringSafeArea(.all)
-            
+
       switch onEnum(of: uiState) {
-        case .initialSeasonsState: LoadingIndicatorView(animate: true)
-        case .seasonDetailsLoaded(let state): SeasonDetailsContent(state)
-        case .seasonDetailsErrorState:
-          FullScreenView(
-            systemName: "exclamationmark.triangle.fill",
-            message: "Something went wrong",
-            buttonText: "Retry",
-            action: { presenter.dispatch(action: ReloadSeasonDetails()) }
-          )
+      case .initialSeasonsState: LoadingIndicatorView(animate: true)
+      case let .seasonDetailsLoaded(state): SeasonDetailsContent(state)
+      case .seasonDetailsErrorState:
+        FullScreenView(
+          systemName: "exclamationmark.triangle.fill",
+          message: "Something went wrong",
+          buttonText: "Retry",
+          action: { presenter.dispatch(action: ReloadSeasonDetails()) }
+        )
       }
     }
     .ignoresSafeArea()
+    .navigationBarTitleDisplayMode(.inline)
+    .navigationBarColor(backgroundColor: .clear)
+    .overlay(
+      VStack(spacing: 0) {
+        GlassToolbar(title: (uiState as? SeasonDetailsLoaded)?.seasonName ?? "", opacity: showGlass)
+        if let state = uiState as? SeasonDetailsLoaded {
+          ProgressView(value: state.watchProgress, total: 1)
+            .progressViewStyle(RoundedRectProgressViewStyle())
+            .background(Color.background)
+            .offset(y: progressViewOffset)
+        }
+      },
+      alignment: .top
+    )
+    .animation(.easeInOut(duration: 0.25), value: showGlass)
+    .edgesIgnoringSafeArea(.top)
     .sheet(isPresented: $showModal) {
       if let uiState = uiState as? SeasonDetailsLoaded {
         ImageGalleryContentView(items: uiState.seasonImages.map { $0.toSwift() })
       }
     }
   }
-    
+
   @ViewBuilder
   private func SeasonDetailsContent(_ state: SeasonDetailsLoaded) -> some View {
     ParallaxView(
-      title: state.seasonName,
-      isRefreshing: state.isUpdating,
       imageHeight: DimensionConstants.imageHeight,
       collapsedImageHeight: DimensionConstants.collapsedImageHeight,
       header: { proxy in
@@ -70,11 +85,9 @@ struct SeasonDetailsView: View {
           headerHeight: proxy.getHeightForHeaderImage(proxy)
         )
       },
-      content: { titleRect in
-
+      content: {
         OverviewBoxView(
-          overview: state.seasonOverview,
-          titleRect: titleRect
+          overview: state.seasonOverview
         )
         .padding()
 
@@ -88,13 +101,18 @@ struct SeasonDetailsView: View {
           onEpisodeHeaderClicked: { presenter.dispatch(action: OnEpisodeHeaderClicked()) },
           onWatchedStateClicked: { presenter.dispatch(action: UpdateSeasonWatchedState()) }
         )
-                
+
         CastListView(casts: toCastsList(state.seasonCast))
       },
-      onBackClicked: {
-        presenter.dispatch(action: SeasonDetailsBackClicked())
-      },
-      onRefreshClicked: {}
+      onScroll: { offset in
+        let opacity = -offset - 150
+        let normalizedOpacity = opacity / 200
+        showGlass = max(0, min(1, normalizedOpacity))
+
+        let startOffset = CGFloat(240)
+        let endOffset = 0
+        progressViewOffset = max(CGFloat(endOffset), startOffset + offset)
+      }
     )
     .onAppear { showModal = state.showSeasonWatchStateDialog }
   }
@@ -113,7 +131,8 @@ struct SeasonDetailsView: View {
             .clear,
             .clear,
             .clear,
-            Color.background.opacity(0.6),
+            .clear,
+            .clear,
             Color.background.opacity(0.8),
             Color.background,
           ]),
@@ -122,7 +141,7 @@ struct SeasonDetailsView: View {
         )
       )
       .frame(height: headerHeight)
-            
+
       ZStack(alignment: .bottom) {
         VStack {
           Spacer()
@@ -135,13 +154,13 @@ struct SeasonDetailsView: View {
               .fontWeight(.regular)
               .foregroundColor(.secondary)
               .alignmentGuide(.view) { d in d[HorizontalAlignment.leading] }
-                        
+
             Text("^[\(state.seasonImages.count) Image](inflect: true)")
               .bodyMediumFont(size: 16)
               .foregroundColor(.textColor)
               .lineLimit(1)
               .alignmentGuide(.view) { d in d[HorizontalAlignment.center] }
-                        
+
             Spacer()
           }
           .padding(16)
@@ -152,22 +171,19 @@ struct SeasonDetailsView: View {
           }
         }
         .frame(height: headerHeight)
-        .opacity(1 - progress)
       }
-            
-      ProgressView(value: state.watchProgress, total: 1)
-        .progressViewStyle(RoundedRectProgressViewStyle())
+      .opacity(1 - progress)
     }
     .frame(height: headerHeight)
     .clipped()
   }
-    
+
   private func toCastsList(_ list: [Cast]) -> [SwiftCast] {
     return list.map { cast -> SwiftCast in
       .init(castId: cast.id, name: cast.name, characterName: cast.characterName, profileUrl: cast.profileUrl)
     }
   }
-    
+
   @ViewBuilder
   private var empty: some View {
     if #available(iOS 17.0, *) {
@@ -189,6 +205,6 @@ struct SeasonDetailsView: View {
 }
 
 private enum DimensionConstants {
-  static let imageHeight: CGFloat = 320
+  static let imageHeight: CGFloat = 350
   static let collapsedImageHeight: CGFloat = 120.0
 }
